Organized files in folders and added login screen with css
