--- Nerdwork/src/application/application.css
@@ -1,1 +1,0 @@
-/* JavaFX CSS - Leave this comment until you have at least create one rule which uses -fx-Property */
