+++ Nerdwork/src/application/functinonality/Professor.java
@@ -1,4 +1,4 @@
-package application;
+package application.functinonality;
 
 import java.util.ArrayList;
