+++ Nerdwork/src/application/functinonality/Subject.java
@@ -1,4 +1,4 @@
-package application;
+package application.functinonality;
 
 import java.util.ArrayList;
