+++ Nerdwork/src/application/gui/classes/Main.java
@@ -1,4 +1,4 @@
-package application;
+package application.gui.classes;
 	
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -12,9 +12,9 @@ public class Main extends Application {
 	@Override
 	public void start(Stage primaryStage) {
 		try {
-			Parent root = FXMLLoader.load(getClass().getResource("MyCourses.fxml"));
+			Parent root = FXMLLoader.load(getClass().getResource("../fxml/Login.fxml"));
 			Scene scene = new Scene(root);
-			scene.getStylesheets().add(getClass().getResource("application.css").toExternalForm());
+//			scene.getStylesheets().add(getClass().getResource("../fxml/generalStyling.css").toExternalForm());
 			primaryStage.setScene(scene);
 			primaryStage.show();
 		} catch(Exception e) {
