+++ Nerdwork/src/application/gui/fxml/Login.fxml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.Separator?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1">
+   <children>
+      <AnchorPane prefHeight="720.0" prefWidth="640.0" style="-fx-background-color: #3889c4;">
+         <children>
+            <Label alignment="CENTER" layoutX="198.0" layoutY="416.0" prefHeight="30.0" prefWidth="200.0" text="An app by students" textFill="WHITE">
+               <font>
+                  <Font name="Segoe UI" size="22.0" />
+               </font>
+            </Label>
+            <Label alignment="CENTER" layoutX="198.0" layoutY="460.0" prefHeight="30.0" prefWidth="200.0" text="For students" textFill="WHITE">
+               <font>
+                  <Font name="Segoe UI" size="16.0" />
+               </font>
+            </Label>
+            <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="223.0" layoutY="235.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../icons/nerd_icon.png" />
+               </image>
+            </ImageView>
+         </children>
+      </AnchorPane>
+      <AnchorPane layoutX="640.0" prefHeight="720.0" prefWidth="640.0">
+         <children>
+            <Pane prefHeight="720.0" prefWidth="640.0" style="-fx-background-color: #fff;">
+               <children>
+                  <PasswordField layoutX="160.0" layoutY="235.0" prefHeight="50.0" prefWidth="300.0" promptText="Password">
+                     <font>
+                        <Font size="14.0" />
+                     </font>
+                  </PasswordField>
+                  <TextField layoutX="160.0" layoutY="150.0" prefHeight="50.0" prefWidth="300.0" promptText="Email">
+                     <font>
+                        <Font name="Segoe UI" size="14.0" />
+                     </font>
+                  </TextField>
+                  <Separator layoutX="160.0" layoutY="364.0" prefHeight="13.0" prefWidth="300.0" />
+                  <Label alignment="CENTER" layoutX="258.0" layoutY="359.0" prefHeight="23.0" prefWidth="90.0" style="-fx-background-color: #fff;" text="Or register as a" textFill="#5b5a5a">
+                     <font>
+                        <Font name="Segoe UI" size="12.0" />
+                     </font>
+                  </Label>
+                  <Separator layoutX="160.0" layoutY="485.0" prefHeight="6.0" prefWidth="300.0" />
+                  <Label alignment="CENTER" layoutX="287.0" layoutY="477.0" prefHeight="23.0" prefWidth="31.0" style="-fx-background-color: #fff;" text="Or a" textFill="#5b5a5a">
+                     <font>
+                        <Font name="Segoe UI" size="12.0" />
+                     </font>
+                  </Label>
+                  <Label alignment="CENTER" layoutX="165.0" layoutY="289.0" prefHeight="23.0" prefWidth="275.0" textFill="#ff6354">
+                     <font>
+                        <Font name="Segoe UI" size="10.0" />
+                     </font>
+                  </Label>
+                  <Label id="forgotPasswordLabel" alignment="CENTER" layoutX="170.0" layoutY="312.0" prefHeight="30.0" prefWidth="120.0" stylesheets="@../css/generalStyling.css" text="Forgot Password" />
+                  <Button id="whiteButton" alignment="CENTER" contentDisplay="CENTER" layoutX="330.0" layoutY="312.0" mnemonicParsing="false" prefHeight="30.0" prefWidth="120.0" stylesheets="@../css/generalStyling.css" text="Sgn In" />
+                  <Button id="whiteButton" layoutX="160.0" layoutY="520.0" mnemonicParsing="false" prefHeight="50.0" prefWidth="300.0" stylesheets="@../css/generalStyling.css" text="Professor">
+                     <font>
+                        <Font name="Segoe UI" size="14.0" />
+                     </font>
+                  </Button>
+                  <Button id="blueButton" layoutX="160.0" layoutY="405.0" mnemonicParsing="false" prefHeight="50.0" prefWidth="300.0" stylesheets="@../css/generalStyling.css" text="Student" textFill="WHITE">
+                     <font>
+                        <Font name="Segoe UI" size="14.0" />
+                     </font>
+                  </Button>
+               </children>
+            </Pane>
+         </children>
+      </AnchorPane>
+   </children>
+</AnchorPane>
