+++ Nerdwork/src/application/gui/icons/house.png
