+++ Nerdwork/src/application/gui/icons/nerd_icon.png
