+++ Nerdwork/src/application/gui/icons/settings.png
